server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
requests>=2.28.0
pyyaml>=6.0
pytest>=7.3.0
pytest-cov>=4.1.0
httpx>=0.24.0
ftw-tools>=1.1.1

# Development dependencies
pre-commit>=3.3.2
ruff>=0.2.0
types-requests>=2.31.0
types-PyYAML>=6.0.12.11
